Going further
Trying to simplify even more the dictionnary so that we only need to
call the function once. No need to set it and call it.  The fct is
already consious of the content. if the key is present it will register
and display it right away. On the second use it will simply display it